Implementations for 1.1.0: Separated constructors for Client & Server (To later implement client-side) Removed SQL dependency over chunk claiming (Now saved to chunk NBT) Removed SQL dependency over player claim options (Now saves to custom NBT) Removed SQL dependency over towns (Now saves to custom NBT) Removed SQL dependency for money (Saves to player NBT.dat) Fixed bed spawns not moving in "/setspawn" command Fixed iron doors from being opened if placed next to a wooden door Fixed death chests only spawning on the surface, now checking for CAVE_AIR Added command to convert from legacy database storage Added saving of custom NBT files to built-in server saving Added chunk position handling for all 16x16x264 blocks in a chunk, for shops Added new wilderness names for nether & end Added a new class for giving and taking money for players, online or offline Added daily reward-money for logging in (Configurable) Added chat message for death chest coordinates (Configurable) Added a version to the config file for porting Added translation keys for money Added separate global/town/local chat Added command for changing chats Added new fun option that passive mobs will seek shelter from rain Modified WitherBoss sound to only be localized

+++ src/main/java/net/TheElm/project/interfaces/ShopSignBlockEntity.java
@@ -44,6 +44,6 @@ public interface ShopSignBlockEntity {
     @Nullable
     Integer getShopItemCount();
     @Nullable
-    Long getShopItemPrice();
+    Integer getShopItemPrice();
     
 }
